requirements.txt
Django==5.1.2
python-dotenv==1.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.9  # Only if you plan to use PostgreSQL later
celery==5.3.6  # If you're using Celery
redis==5.0.1  # If you're using Redis with Celery

asgiref==3.9.1
coloredlogs==15.0.1
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
flatbuffers==25.2.10
humanfriendly==10.0
mpmath==1.3.0
numpy==2.2.6
onnxruntime==1.22.1
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
pip==24.0
protobuf==6.32.0
PyJWT==2.10.1
pyreadline3==3.5.4
sqlparse==0.5.3
sympy==1.14.0
tzdata==2025.2